separate InstrumentationContext and MockGetter
